requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
httpx==0.27.2
python-dotenv==1.0.1
pyngrok==7.2.0
python-nmap==0.7.1
